jeff refactoring edits
